feat: status page over websocket

+++ src/test/kotlin/pl/folsoft/elevator/model/LiftTest.kt
@@ -4,7 +4,6 @@ import io.kotlintest.data.suspend.forall
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.StringSpec
 import io.kotlintest.tables.row
-import java.util.concurrent.atomic.AtomicInteger
 
 internal class LiftTest : StringSpec() {
 
@@ -12,17 +11,17 @@ internal class LiftTest : StringSpec() {
 
         "calculating floor distance" {
             forall(
-                    row(1, 2, 1),
-                    row(2, 1, 1),
-                    row(-1, -2, 1),
-                    row(-2, -1, 1),
-                    row(-1, 2, 3),
-                    row(2, -1, 3),
-                    row(1, 1, 0),
-                    row(0, 0, 0),
-                    row(-1, -1, 0)
+                row(1, 2, 1),
+                row(2, 1, 1),
+                row(-1, -2, 1),
+                row(-2, -1, 1),
+                row(-1, 2, 3),
+                row(2, -1, 3),
+                row(1, 1, 0),
+                row(0, 0, 0),
+                row(-1, -1, 0)
             ) { currentFloor, addressFloor, expectedDistance ->
-                AtomicInteger(currentFloor).distance(addressFloor) shouldBe expectedDistance
+                currentFloor.distance(addressFloor) shouldBe expectedDistance
             }
         }
     }
